WIP: Rendering componentIds to the GeomData buffer are working now.

+++ src/Renderer/Shaders/GLSL/drawItemId.glsl
@@ -11,8 +11,8 @@ int getDrawItemId() {
   return drawId;
 }
 
-ivec3 getDrawItemIds() {
-  return ivec3(drawId, 0, -1);
+vec3 getDrawItemIds() {
+  return vec3(drawId, 0, -1);
 }
 
 #else // EMULATE_MULTI_DRAW
@@ -23,11 +23,11 @@ int getDrawItemId() {
   return int(texelFetch(drawIdsTexture, drawIdsArrayCoords, 0).r + 0.5);
 }
 
-ivec3 getDrawItemIds() {
+vec3 getDrawItemIds() {
   ivec2 drawIdsTextureSize = textureSize(drawIdsTexture, 0);
   ivec2 drawIdsArrayCoords = ivec2(gl_DrawID % drawIdsTextureSize.x, gl_DrawID / drawIdsTextureSize.x);
   vec4 color = texelFetch(drawIdsTexture, drawIdsArrayCoords, 0);
-  return ivec3(int(color.r + 0.5), int(color.g + 0.5), int(color.b));
+  return vec3(int(color.r + 0.5), int(color.g + 0.5), int(color.b));
 }
 
 #endif // EMULATE_MULTI_DRAW
@@ -50,12 +50,12 @@ int getDrawItemId() {
   }
 }
 
-ivec3 getDrawItemIds() {
+vec3 getDrawItemIds() {
   if (instancedDraw == 0) {
-    return ivec3(drawItemId, 0, -1);
+    return vec3(drawItemId, 0, -1);
   }
   else {
-    return ivec3(instancedIds, 0, -1);
+    return vec3(instancedIds, 0, -1);
   }
 }
 
@@ -65,8 +65,8 @@ int getDrawItemId() {
   return drawItemId;
 }
 
-ivec3 getDrawItemIds() {
-  return ivec3(drawItemId, 0, -1);
+vec3 getDrawItemIds() {
+  return vec3(drawItemId, 0, -1);
 }
 
 #endif // ENABLE_FLOAT_TEXTURES
